pyproject.toml
[tool.poetry]
name = "bbc-to-spotify"
version = "0.1.0"
description = ""
authors = ["gietema <jochem@giete.ma>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
spotify = "^0.10.2"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"
spotipy = "^2.23.0"
click = "^8.1.7"
python-levenshtein = "^0.23.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
